Plugins: categorize linux.malfind as a malware plugin
