+++ test/plugins/linux/linux.py
@@ -251,7 +251,7 @@ class TestLinuxKthreads:
 class TestLinuxMalfind:
     def test_linux_generic_malfind(self, image, volatility, python):
         rc, out, _err = test_volatility.runvol_plugin(
-            "linux.malfind.Malfind", image, volatility, python
+            "linux.malware.malfind.Malfind", image, volatility, python
         )
 
         # linux-sample-1.bin has no process memory ranges with potential injected code.
@@ -479,7 +479,10 @@ class TestLinuxIomem:
 class TestLinuxKeyboardNotifiers:
     def test_linux_generic_keyboard_notifiers(self, image, volatility, python):
         rc, out, _err = test_volatility.runvol_plugin(
-            "linux.malware.keyboard_notifiers.Keyboard_notifiers", image, volatility, python
+            "linux.malware.keyboard_notifiers.Keyboard_notifiers",
+            image,
+            volatility,
+            python,
         )
 
         # linux-sample-1.bin has no suspicious results for this plugin.
